+++ fluent-kafka-streams-tests/src/test/java/com/bakdata/fluent_kafka_streams_tests/test_applications/MirrorAvroNonDefaultSerde.java
@@ -1,0 +1,78 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2025 bakdata
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package com.bakdata.fluent_kafka_streams_tests.test_applications;
+
+import com.bakdata.kafka.Preconfigured;
+import io.confluent.kafka.serializers.AbstractKafkaSchemaSerDeConfig;
+import io.confluent.kafka.streams.serdes.avro.SpecificAvroSerde;
+import java.util.HashMap;
+import java.util.Map;
+import lombok.experimental.UtilityClass;
+import org.apache.avro.specific.SpecificRecord;
+import org.apache.kafka.common.serialization.Serde;
+import org.apache.kafka.common.serialization.Serdes.StringSerde;
+import org.apache.kafka.streams.StreamsBuilder;
+import org.apache.kafka.streams.StreamsConfig;
+import org.apache.kafka.streams.Topology;
+import org.apache.kafka.streams.kstream.Consumed;
+import org.apache.kafka.streams.kstream.KStream;
+import org.apache.kafka.streams.kstream.Produced;
+
+@UtilityClass
+public class MirrorAvroNonDefaultSerde {
+    private static final String INPUT_TOPIC = "input";
+    private static final String OUTPUT_TOPIC = "output";
+
+    public static Map<String, Object> getKafkaProperties() {
+        final String brokers = "localhost:9092";
+        final Map<String, Object> kafkaConfig = new HashMap<>();
+        kafkaConfig.put(StreamsConfig.APPLICATION_ID_CONFIG, "mirror");
+        kafkaConfig.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, brokers);
+        kafkaConfig.put(AbstractKafkaSchemaSerDeConfig.SCHEMA_REGISTRY_URL_CONFIG, "mock://");
+        kafkaConfig.put(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, StringSerde.class);
+        kafkaConfig.put(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, StringSerde.class);
+        return kafkaConfig;
+    }
+
+    public static Topology getTopology() {
+        final StreamsBuilder builder = new StreamsBuilder();
+        final Preconfigured<Serde<SpecificRecord>> keySerde = getKeySerde();
+        final Preconfigured<Serde<SpecificRecord>> serde = getValueSerde();
+        final KStream<SpecificRecord, SpecificRecord> input = builder.stream(INPUT_TOPIC,
+                Consumed.with(keySerde.configureForKeys(getKafkaProperties()),
+                        serde.configureForValues(getKafkaProperties())));
+        input.to(OUTPUT_TOPIC, Produced.with(keySerde.configureForKeys(getKafkaProperties()),
+                serde.configureForValues(getKafkaProperties())));
+        return builder.build();
+    }
+
+    public static Preconfigured<Serde<SpecificRecord>> getKeySerde() {
+        return Preconfigured.create(new SpecificAvroSerde<>());
+    }
+
+    public static Preconfigured<Serde<SpecificRecord>> getValueSerde() {
+        return Preconfigured.create(new SpecificAvroSerde<>());
+    }
+}